Fixed bugs+REFORMAT type of code

+++ src/main/java/org/project/dao/CityDAO.java
@@ -2,6 +2,6 @@ package org.project.dao;
 
 import org.project.models.City;
 
-public interface CityDAO extends GenericDAO<City, Integer> {
+public interface CityDAO extends GenericDAO<City, Long> {
     // additional methods related to City can be added here
 }
