+++ src/main/java/org/project/dao/CountryDAO.java
@@ -2,6 +2,6 @@ package org.project.dao;
 
 import org.project.models.Country;
 
-public interface CountryDAO extends GenericDAO<Country, Integer> {
+public interface CountryDAO extends GenericDAO<Country, Long> {
     // additional methods related to Country can be added here
 }
