+++ src/main/java/org/project/dao/UserDAO.java
@@ -2,6 +2,6 @@ package org.project.dao;
 
 import org.project.models.User;
 
-public interface UserDAO extends GenericDAO<User, Integer> {
+public interface UserDAO extends GenericDAO<User, Long> {
     // additional methods related to Country can be added here
 }
