+++ src/main/java/org/project/dao/implementation/UserDAOImpl.java
@@ -6,7 +6,7 @@ import org.project.models.User;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class UserDAOImpl extends GenericDAOImpl<User, Integer> implements UserDAO {
+public class UserDAOImpl extends GenericDAOImpl<User, Long> implements UserDAO {
 
     public UserDAOImpl(SessionFactory sessionFactory) {
         super(sessionFactory);
