+++ src/main/java/org/project/service/HotelService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface HotelService {
     List<Hotel> getAllHotels();
 
-    Hotel getHotelById(int id);
+    Hotel getHotelById(long id);
 
     void saveHotel(Hotel hotel);
